GFA1: Clarify that names are unique (#87)
Currently the specification does not clearly state that names are unique in a GFA1 file. Although in most implementations segment names are unique, this is not true for path names, which are sometimes equal to existing segment names. However, creates problems in the conversion to GFA2 (which requires IDs to be unique, with the exception of external sequences in fragments). Furthermore not having unique identifiers have other disadvantages: e.g. it breaks generic functions, which return a line of any kind from an ID, and does not allow to store named lines in dictionary-like structures.