merge v1.1 with v1.0
v1.1 exists only to define W-lines.  Most tools already ignore lines in
namespaces that they don't handle.  By making this explicit in the v1.0
specification, the need for v1.1 is obviated.

This limits user confusion and avoids needless fragmentation of the GFA
ecosystem.

Additionally, an explanation of the various pangenome representation
approaches has been added to the introduction.