added gfastats to the list of tools supporting gfa2 and proposed specification for cuts in segments
As discussed today with @lh3, I am proposing to allow optional specification of subsequences directly in a path component, e.g. e.g. ptg000124l.1(397101:505533)+

We have devised a new tool (gfastats) that can take an agp file, for example from a scaffolding tool, and overlay the information on the original graph. When cuts are made within contigs, the segment is unmutable but we represent this event with coordinates associated to the component in the path, allowing to easily identify where cuts have occurred in the original segments